Dimension

Spindles.dim — Function

dim(p::Polytope)

Compute the dimension of p.

This is done by computing the length of a maximal chain in the face lattice of p, i.e., a finite sequence of faces

\[\emptyset = F_{-1} \subsetneq F_0 \subsetneq F_1 \subsetneq \dots \subsetneq F_d\]

for which $d$ is maximal among all such sequences. Then $F_d$ must be p itself, and $d$ is its dimension.

See also Polyhedra.dim.

Examples

julia> p = Polytope([0 0; 1 0; 0 1; 1 1])
Polytope{Rational{BigInt}}

julia> dim(p)
2

dim(p::Polytope, indices)
dim(p::Polytope, i::Int)

Compute the dimension of the face of p that is defined by the inequalities in the collection indices, or by the single inequality at index i. If indices is empty, this is the same as dim(p).

The implementation idea is the same as above except that the maximal face of the chain is the face defined by indices.

See also codim.

Examples

julia> p = Polytope([-1 0; 1 0; 0 -1; 0 1], [0, 1, 0, 1])
Polytope{Rational{BigInt}}

julia> dim(p, Int[])
2

julia> dim(p, 1)
1

julia> dim(p, [1, 3])
0

julia> dim(p, [1, 2])
-1

dim can be used to test whether an inequality is facet-defining for a polytope p (see the example above). In high dimensions, this may be slow, since dim constructs a chain of faces between the empty face and the given face – which in the case of a facet amounts to a maximal chain except for the maximal face p. In this case, however, the complementary approach is fast: Find a chain of faces not between the empty face and the given face, but between the face and the polytope p itself. From the length of the resulting chain, one can easily read off the codimension of the face (which equals 1 for a facet):

Spindles.codim — Function

codim(p::Polytope, indices)
codim(p::Polytope, i::Int)

Compute the codimension dim(p) - dim(p, indices) of the face of p that is defined by the inequalities in indices, or by the single inequality at index i.

For the sake of consistency, the codimension of the empty face of a $d$-dimensional polytope is defined as $d+1$. The implementation is complementary to dim and computes a maximal chain of faces between the given face and p itself in the face lattice of p.

See also dim.

Examples

julia> p = Polytope([-1 0; 1 0; 0 -1; 0 1], [0, 1, 0, 1])
Polytope{Rational{BigInt}}

julia> codim(p, Int[])
0

julia> codim(p, 1)
1

julia> codim(p, [1, 3])
2

julia> codim(p, [1, 2])
3