Faces

Enumeration

Spindles.graph — Function

graph(s::Spindle)

Return the graph (aka 1-skeleton) of s, which is a simple undirected graph of type Graphs.SimpleGraphs.SimpleGraph.

Spindles.jl implements an algorithm for enumerating all faces of a given dimension. On large inputs, this algorithm works best for near-simple spindles (with few degenerate vertices).

Spindles.facesofdim — Function

facesofdim(s::Spindle, k::Int)

Enumerate all faces of dimension k of the spindle s, each one given by a vector containing the indices of all incident halfspaces/facets (starting at 1). Return them in one vector.

Note

difference from Polyhedra.Index

Recursive bottom-up computation. Most (memory-)efficient for near-simple polytopes, for which faces are contained in few facets each.

Spindles.nfacesofdim — Function

nfacesofdim(s::Spindle, k::Int)

Count the k-dimensional faces of the spindle s. ... Equivalent with length(facesofdim(s, k)) if not nothing. Uses the convention that the dimension of the empty face is -1.

Length of paths

Spindles.dist — Function

dist(s::Spindle, apex::Int, v::Int)

Compute the distance between apex and vertex v in the graph of spindle s.

Note

Results are affected by resetting the apices with setapex!.

Examples

julia> square = Spindle([1 0; 0 1; -1 0; 0 -1], [1, 1, 1, 1]);

julia> apx1, apx2 = setapex!(square, 1)
2-element Vector{Int64}:
 1
 4

julia> dist(square, apx1, 4)
2

julia> apx1, apx2 = setapex!(square, 2)
2-element Vector{Int64}:
 2
 3

julia> dist(square, apx1, 4)
1

Good faces

Spindles.FaceState — Type

FaceState

Fields

good::Bool

facets: all incident facets

edges

vsets

Spindles.isgood2face — Function

isgood2face(s::Spindle, facets)

Test the face defined by facets for being a good 2-face of the spindle s. Return a FaceState.

(false, nothing, nothing) if not a good 2-face.