Good 2-faces

Spindles.FaceState — Type

FaceState

Fields

good::Bool

indices: indices of all tight inequalities for the face

edges

vsets

Spindles.isgood2face — Function

isgood2face(p::Polytope, indices, src, dst)

Test the face defined by indices for being a good 2-face of the polytope p with respect to the two vertices src and dst. Return a FaceState.

See this tutorial for an informal explanation of what it means for a 2-face to be good.