File I/O

Spindles.readineq — Function

readineq(filename, T [, comment_char])

File format: labels b -A Return A, b, labels where A is a matrix of type Rational{T}.

Lines starting with a comment_char character (default is '#') and all characters on a line following a comment_char are ignored.

Examples

julia> str = "# unit square\na 1 -1 0\nb 1 0 -1\nc 1 1 0\nd 1 0 1";

julia> println(str)
# unit square
a 1 -1 0
b 1 0 -1
c 1 1 0
d 1 0 1

julia> open("square.txt", "w") do io
           write(io, str)
       end;

julia> A, b, labels = readineq("square.txt", Int);

julia> A
4×2 Matrix{Rational{Int64}}:
  1   0
  0   1
 -1   0
  0  -1

julia> b
4-element Vector{Rational{Int64}}:
 1
 1
 1
 1

julia> labels
4-element Vector{String}:
 "a"
 "b"
 "c"
 "d"

julia> rm("square.txt")

Spindles.writeineq — Function

writeineq(outfilename, A::AbstractMatrix, b::AbstractVector [, labels, labels_plusminus, comments, comment_char])

Write the inequality description $Ax \le b$ to outfilename.

The file format is [labels b -A]. First column contains inequality labels.

Keywords

labels: If not specified, use inequality indices.

labels_plusminus::Bool: Defaults to false

comments: Each element in comments will be printed on its own line, starting with comment_char and a whitespace. Possible internal line breaks are ignored.

comment_char: default is '#'

Spindles.readrational — Function

readrational(filename, T)

T type of numerator and denominator