Plotting faces

Since the initial purpose of Spindles.jl is the detection of special 2-faces, the package provides a function to visualize those.

Implementation note

The implementation in Spindles.jl relies on Plots.jl. Polygons such as 2-faces of polytopes may also be visualized, more generally, using Polyhedra.jl.

The plotting function of Spindles.jl is called plot2face and is best described using an example.

Example

We will use the following simple polytope, a square in 2D given by

\[\begin{aligned} -1 \le x_1 &\le 1 \\ -1 \le x_2 &\le 1 \end{aligned}\]

square = Polytope([1 0; 0 1; -1 0; 0 -1], [1, 1, 1, 1])
vertices(square)

4-element iterator of Vector{Rational{BigInt}}:
 Rational{BigInt}[-1, -1]
 Rational{BigInt}[1, -1]
 Rational{BigInt}[-1, 1]
 Rational{BigInt}[1, 1]

Since square is two-dimensional already, the most basic form of calling plot2face just takes as arguments the polytope and a list of inequality indices that define the face to be drawn. In our example, this list is empty to get the maximal face, the polytope itself:

plot2face(square, Int[])

By default, vertices are labeled by their index in vertices(square). The edge labels indicate which facet-defining inequalities for square define the edge. We may relabel vertices and edges using the keyword arguments vertexlabels and ineqlabels as follows:

plot2face(square, Int[]; vertexlabels=["a", "b", "c", "d"], ineqlabels=["^", "*", "+", "-"])

If we are only interested in the combinatorics of square (which of course isn't all too interesting), we can tell plot2face to make a plot of the graph of square, where vertices are placed equidistantly on a cycle. In our example, this of course does not change too much:

plot2face(square, Int[]; usecoordinates=false)

Another option that has its origins in the theoretical analysis of 2-faces allows for marking up a pair of edges. To see the effect, let us drop the edge labels for the moment. This is achieved by setting ineqlabels to nothing.

plot2face(square, Int[]; ineqlabels=nothing, directed_edges=([1,3], [2,4]))

Now perturb the square slightly so that those two edges are no longer parallel.

perturbed_square = Polytope([1 -1//8; 0 1; -1 -1//8; 0 -1], [9//8, 1, 9//8, 1])
vertices(perturbed_square)

4-element iterator of Vector{Rational{Int64}}:
 Rational{Int64}[-1, -1]
 Rational{Int64}[1, -1]
 Rational{Int64}[-5//4, 1]
 Rational{Int64}[5//4, 1]

plot2face(perturbed_square, Int[]; ineqlabels=nothing, directed_edges=([1,3], [2,4]))

Full reference

Spindles.plot2face — Function

plot2face(
    p::Polytope, indices;
    usecoordinates = true,
    vertexlabels = ,
    ineqlabels = ,
    directed_edges = nothing,
    # plot kw args
    kw...
)

Make a plot of the 2-face of p that is defined by the inequalities in indices, either as a 2D projection onto the plane (if the argument usecoordinates is set to true) or as a (combinatorial) plot of its graph otherwise.

Keywords

usecoordinates: If true (default), plot a 2D projection. Otherwise draw the graph.

vertexlabels: An indexable collection (such as AbstractVector or AbstractDict) of strings, or nothing. If nothing, no vertex labels are shown. Otherwise, the label of vertex i will be vertexlabels[i], where missing values are treated as "". If vertexlabels is unspecified, use vertex indices as default labels.

ineqlabels: A collection of strings to be used as facet labels, or nothing to suppress labels. If unspecified, use inequality indices as default labels.

directed_edges: A tuple of edges ([s,t], [u,v]) that are drawn as directed edges.

The remaining keyword arguments kw... are passed to plot and can be any plot, subplot, or axis attributes. See also the Plots.jl documentation pages for a list of available attributes. Some of them are used by plot2face with a different default value than in Plots.jl. Notable keyword arguments among those are:

size: A tuple of Integers that determines the plot size. Defaults here to (300,300).

aspect_ratio: Defaults to :equal if usecoordinates is false, and :auto otherwise (default in Plots.jl).

The default behaviour can be overwritten by explicitly passing new values as keyword arguments in kw... to plot2face. Anything in kw... takes precedence over the default behaviour in plot2face, except for (most) attributes related to annotations. They are hardcoded in plot2face.