Polytopes

The main type defined by Spindles.jl is the type Polytope that represents a polytope. Objects of this type are constructed either from a V-representation (a list of points whose convex hull is the polytope) or from an H-representation (a system of linear inequalities whose set of solutions is the polytope):

Spindles.Polytope — Type

Polytope(V::AbstractVector{AbstractVector})

Create a polytope from the convex hull of the collection of points V.

Polytope(V::AbstractMatrix)

Create a polytope from the convex hull of the rows of V.

Polytope(A::AbstractMatrix, b::AbstractVector)

Create a polytope from its H-representation $Ax \le b$. If the polyhedron defined by $Ax \le b$ is unbounded, throw an error.

Examples

julia> p = Polytope([[0, 0], [1, 0], [0, 1], [1, 1]])
Polytope{Rational{BigInt}}

creates the polytope with vertices $(0,0),(1,0),(0,1)$, and $(1,1)$ (the two-dimensional 0/1 cube). This is the same as

julia> q = Polytope([0 0; 1 0; 0 1; 1 1])
Polytope{Rational{BigInt}}

julia> p == q
true

Note

Two Polytopes are considered equal by the == operator if and only if they have the same set of vertices.

Equivalently, our 2D polytope may be created from the inequality description

\[\begin{aligned} 0 \le x_1 &\le 1 \\ 0 \le x_2 &\le 1 \end{aligned}\]

which translates to

julia> r = Polytope([-1 0; 1 0; 0 -1; 0 1], [0, 1, 0, 1])
Polytope{Rational{BigInt}}

julia> p == r
true

Unlike for the first two constructors above, there is no guarantee that a polyhedron defined by a general system $Ax \le b$ is bounded (and, hence, a polytope). Indeed, if we drop any of the four inequalities above – say the last one –, this property is lost:

julia> Polytope([-1 0; 1 0; 0 -1], [0, 1, 0])
ERROR: ArgumentError: got an unbounded polyhedron
[...]

Type parameter

Polytope is a parametric type. Namely, the precise type of each of the three objects constructed above is Polytope{Rational{BigInt}}, where the parameter Rational{BigInt} is called the element type and is inferred from the type of the data passed to the constructor. For example, Rational{BigInt} indicates that Polytope uses exact rational arithmetic to store and manipulate the coefficients in a V- or H-representation. For integer data, this is the default choice.

Suppose that we change some of the input data above to floating-point numbers, such as

julia> Polytope([-1.0 0.0; 1 0; 0 -1; 0 1], [0, 1, 0, 1])
Polytope{Float64}

Then we get a different element type, namely Float64.

The type of arithmetic used to represent and manipulate the Polytope object can also be set using an optional constructor argument that is borrowed from Polyhedra.jl (and that may be passed to any of the Polytope constructors):

julia> using Polyhedra: DefaultLibrary

julia> Polytope([-1.0 0.0; 1 0; 0 -1; 0 1], [0, 1, 0, 1], DefaultLibrary{Rational{BigInt}}())
Polytope{Rational{BigInt}}

This argument now specifies a library for polyhedral computations (the "backend" of Polyhedra.jl) that is to be used for the internal representation of the Polytope object. In this case, we chose the default library implemented in Polyhedra.jl but forced rational data instead of Floats.

See also the JuliaPolyhedra website for a list of all supported libraries. For example, to use CDDLib with exact rational arithmetic, do

import CDDLib
Polytope(A, b, CDDLib.Library(:exact))

Also note that lists of points or inequality descriptions passed to one of the Polytope constructors need not be minimal: They may include non-vertices; and redundant inequalities or implicit equations in a system of linear inequalities are allowed, too. However, these can be detected, see the section on Redundancy and implicit equations below.

Vertices

Spindles.vertices — Function

vertices(p::Polytope)

Return an iterator over the vertices of the polytope p.

Spindles.nvertices — Function

nvertices(p::Polytope)

Count the vertices of p.

Incidence

Spindles.incidentvertices — Function

incidentvertices(p::Polytope, indices::AbstractVector{Int})

List the indices of all vertices of the polytope p for which each inequality in indices is tight.

If indices is empty, this is the same as collect(1:nvertices(p)).

Redundancy and implicit equations

When constructing a Polytope from an inequality description, any inequality description is allowed. In particular, it may contain redundant inequalities (whose deletion leaves the polytope unchanged) or implicit equations (inequalities that are satisfied at equality by all points in the polytope). So the inequalities in a given description may be partitioned into three sets:

a minimal set of facet-defining inequalities (which may not be unique),

a (possibly empty) set of implicit equations contained in the inequality system, and

all remaining inequalities (that may safely be deleted).

This partition can be computed using the functions facets and impliciteqs, which return the first two sets of the partition.

Spindles.facets — Function

facets(p::Polytope)

Return a minimal subset of inequality indices that contains one inequality for each facet of p.

If multiple inequalities in the description of p define the same facet, the one with the smallest index is selected.

See also nfacets, impliciteqs.

Examples

Spindles.nfacets — Function

nfacets(p::Polytope)

Count the facets of p. Shorthand for length(facets(p)).

See also nfacesofdim, facets.

Spindles.impliciteqs — Function

impliciteqs(p::Polytope)

Return the indices of all inequalities that are implicit equations, i.e., that are satisfied at equality for each point in p.

See also facets.

Spindles

Spindles.apices — Function

apices(p::Polytope [, apex::Int]; checkredund=true)

Check whether p is a spindle, i.e., if there is a pair of vertices (the apices) for which each facet of p is incident to exactly one of them. If p has a pair of apices, return their indices; otherwise return nothing.

The optional argument apex specifies the index of a vertex that is to be taken as one of the apices.

Keywords

checkredund::Bool: If true (default), first detect all inequalities in the description of p that are not facet-defining for p. Especially in higher dimensions, this step may be time- and memory-consuming. To speed up computations for polytopes created from inequality descriptions that are known to be minimal, set checkredund to false.

Note

For polytopes created from systems that include non-facet-defining inequalities, disabling checkredund can only produce false negatives. All vertices returned by apices are guaranteed to be apices.

Examples

julia> square = Polytope([1 0; 0 1; -1 0; 0 -1], [1, 1, 1, 1]);

julia> vertices(square)
4-element iterator of Vector{Rational{BigInt}}:
 Rational{BigInt}[-1, -1]
 Rational{BigInt}[1, -1]
 Rational{BigInt}[-1, 1]
 Rational{BigInt}[1, 1]

julia> apices(square)
2-element Vector{Int64}:
 1
 4

julia> apices(square, 2)
2-element Vector{Int64}:
 2
 3