Combinatorics

Dimension

Spindles.Polytopes.dim — Function

dim(p::Polytope)

Compute the dimension of p.

This is done by computing the length of a maximal chain in the face lattice of p, i.e., a finite sequence of faces

\[\emptyset = F_{-1} \subsetneq F_0 \subsetneq F_1 \subsetneq \dots \subsetneq F_d\]

for which $d$ is maximal among all such sequences. Then $F_d$ must be p itself, and $d$ is its dimension.

See also Polyhedra.dim.

Examples

julia> p = Polytope([0 0; 1 0; 0 1; 1 1])
Polytope{Rational{BigInt}}

julia> dim(p)
2

dim(p::Polytope, indices)
dim(p::Polytope, i::Int)

Compute the dimension of the face of p that is defined by the inequalities in the collection indices, or by the single inequality at index i. If indices is empty, this is the same as dim(p).

The implementation idea is the same as above except that the maximal face of the chain is the face defined by indices.

See also codim.

Examples

julia> p = Polytope([-1 0; 1 0; 0 -1; 0 1], [0, 1, 0, 1])
Polytope{Rational{BigInt}}

julia> dim(p, Int[])
2

julia> dim(p, 1)
1

julia> dim(p, [1, 3])
0

julia> dim(p, [1, 2])
-1

dim can be used to test whether an inequality is facet-defining for a polytope p (see the example above). In high dimensions, this may be slow, since dim constructs a chain of faces between the empty face and the given face – which in the case of a facet amounts to a maximal chain except for the maximal face p. In this case, however, the complementary approach is fast: Find a chain of faces not between the empty face and the given face, but between the face and the polytope p itself. From the length of the resulting chain, one can easily read off the codimension of the face (which equals 1 for a facet):

Spindles.Polytopes.codim — Function

codim(p::Polytope, indices)
codim(p::Polytope, i::Int)

Compute the codimension dim(p) - dim(p, indices) of the face of p that is defined by the inequalities in indices, or by the single inequality at index i.

For the sake of consistency, the codimension of the empty face of a $d$-dimensional polytope is defined as $d+1$. The implementation is complementary to dim and computes a maximal chain of faces between the given face and p itself in the face lattice of p.

See also dim.

Examples

julia> p = Polytope([-1 0; 1 0; 0 -1; 0 1], [0, 1, 0, 1])
Polytope{Rational{BigInt}}

julia> codim(p, Int[])
0

julia> codim(p, 1)
1

julia> codim(p, [1, 3])
2

julia> codim(p, [1, 2])
3

Face enumeration

Spindles.jl implements an algorithm for enumerating all faces of a given dimension. The algorithm is optimized for near-simple polytopes, i.e., polytopes with few degenerate vertices. It is inspired by an algorithm for computing the graph of a polytope described here (see also the paper).

Spindles.Polytopes.facesofdim — Function

facesofdim(p::Polytope, k::Int)

Enumerate all faces of dimension k of the polytope p. Each face is given by a list of the indices of its incident halfspaces.

Note here that the empty face $\emptyset$ (which is the unique face of dimension −1 by convention) is given by the list of all halfspace indices, as the intersection of all facets of a polytope is empty.

Difference from Polyhedra.jl

The index of a halfspace is the index of the corresponding inequality in the linear description of p, where (explicitly given) equality constraints (defining hyperplanes) are ignored. This is different from the way that indices are treated in Polyhedra.jl, where hyperplanes and halfspaces share the same set of indices.

Note

The algorithm proceeds recursively and computes faces bottom-up, starting from the vertices. The results are cached internally in the Polytope object p. Therefore, subsequent calls to facesofdim(p, l) for any $l \le k$ do not cost anything.

See also nfacesofdim, dim.

Spindles.Polytopes.nfacesofdim — Function

nfacesofdim(p::Polytope, k::Int)

Count the k-dimensional faces of the polytope p. Shorthand for length(facesofdim(p, k)).

See also nfacets, facesofdim, dim.

Examples

julia> V = [(i >> j) & 1 for i=0:7, j=0:2]
8×3 Matrix{Int64}:
 0  0  0
 1  0  0
 0  1  0
 1  1  0
 0  0  1
 1  0  1
 0  1  1
 1  1  1

julia> cube = Polytope(V);

julia> nfacesofdim.(cube, -1:3)
5-element Vector{Int64}:
  1
  8
 12
  6
  1

Graphs and distances

For example, calling facesofdim(p, 1) lists the incident halfspaces of each edge of the polytope p. These edges, given as pairs of adjacent vertices, define the graph (or 1-skeleton) of p. The graph can also be retrieved directly:

Spindles.Polytopes.graph — Function

graph(p::Polytope)

Return the graph of the polytope p, which is a simple undirected graph of type Graphs.SimpleGraphs.SimpleGraph.

Spindles.Polytopes.dist — Function

dist(p::Polytope, u::Int, v::Int)

Compute the distance between vertices u and v in the graph of p.

Examples

julia> p = Polytope([0 0; 0 1; 1 1; 1 0]);

julia> vertices(p)
4-element iterator of Vector{Rational{BigInt}}:
 Rational{BigInt}[0, 0]
 Rational{BigInt}[0, 1]
 Rational{BigInt}[1, 1]
 Rational{BigInt}[1, 0]

julia> dist(p, 1, 3)
2