Generators for standard polytopes

Spindles.Polytopes.crosspolytope — Method

crosspolytope(n)

Create the n-dimensional standard cross-polytope whose vertices are the $2n$ positive and negative standard basis vectors in $\mathbb{R}^n$.

Examples

julia> collect(vertices(crosspolytope(3)))
6-element Vector{Vector{Rational{BigInt}}}:
 [1, 0, 0]
 [-1, 0, 0]
 [0, 1, 0]
 [0, -1, 0]
 [0, 0, 1]
 [0, 0, -1]

Spindles.Polytopes.cube — Method

cube(n)

Create the n-dimensional standard hypercube $[-1,1]^n$.

Examples

julia> collect(vertices(cube(2)))
4-element Vector{Vector{Rational{BigInt}}}:
 [-1, -1]
 [1, -1]
 [-1, 1]
 [1, 1]

Spindles.Polytopes.permutahedron — Method

permutahedron(n)

Create the n-th permutahedron. It is defined as the convex hull of the vectors $(\pi(1), \pi(2), \dots, \pi(n))$ for all permutations $\pi$ of $\{1,2, \dots, n\}$.

Examples

julia> p = permutahedron(3);

julia> collect(vertices(p))
6-element Vector{Vector{Rational{Int64}}}:
 [3, 2, 1]
 [3, 1, 2]
 [2, 3, 1]
 [1, 3, 2]
 [2, 1, 3]
 [1, 2, 3]

julia> nvertices(p) == factorial(3)
true

Note that the dimension of the $n$-th permutahedron is $n-1$:

julia> dim(p)
2

Spindles.Polytopes.simplex — Method

simplex(n)

Create the n-dimensional simplex whose vertices are the standard basis vectors in $\mathbb{R}^n$ plus the origin.

Examples

julia> collect(vertices(simplex(3)))
4-element Vector{Vector{Rational{BigInt}}}:
 [1, 0, 0]
 [0, 1, 0]
 [0, 0, 1]
 [0, 0, 0]