Operations on polytopes

Equality and containment

Base.:== — Function

==(p::Polytope, q::Polytope)

Check whether polytopes p and q have the same set of vertices.

Examples

julia> Polytope([1 0; 0 1]) == Polytope([1 0; 0 1; 0 1; 1//2 1//2])
true

julia> Polytope([1 0; 0 1]) == Polytope([1.0 -0.0; 0 1])
true

julia> Polytope([1 0; 0 1]) == Polytope([0.999999 0; 0 1])
false

julia> p = Polytope([-1 0; 0 -1; 2 1], [0, 0, 3]);

julia> p == Polytope(collect(vertices(p)))
true

Base.in — Function

in(x, p)

Check whether the vector x is in the polytope p.

Examples

julia> p = Polytope([0 0; 1 0; 0 1]);

julia> [1, 1] in p
false

julia> (sum(vertices(p)) / nvertices(p)) in p
true

Constructions

The following operations on Polytopes are supported:

Cartesian product

Minkowski sum

polarization

rescalation and translation

To define more complex operations that act on the set of vertices, use map. See below for an example of how projections can be defined using map. More examples can be found in this tutorial.

Base.:* — Method

*(p::Polytope, q::Polytope)

Cartesian product of polytopes p and q.

Examples

A cube is the product of line segments:

julia> p = Polytope([[0], [1]]);

julia> collect(vertices(p * p * p))
8-element Vector{Vector{Rational{BigInt}}}:
 [1, 1, 1]
 [1, 1, 0]
 [1, 0, 1]
 [1, 0, 0]
 [0, 1, 1]
 [0, 1, 0]
 [0, 0, 1]
 [0, 0, 0]

Base.:* — Method

*(δ::Real, p::Polytope)
*(p::Polytope, δ::Real)

Rescale p by the scalar factor δ. Returns a new Polytope.

See also map.

Examples

julia> p = Polytope([[0, 0], [1, 0], [0, 1]])
Polytope{Rational{BigInt}}

julia> collect(vertices(p * 2))
3-element Vector{Vector{Rational{BigInt}}}:
 [0, 0]
 [2, 0]
 [0, 2]

julia> 2p == p + p
true

Note that the element type of the rescaled polytope may be different from that of p:

julia> 1//2 * p
Polytope{Rational{BigInt}}

julia> 0.5 * p
Polytope{BigFloat}

julia> 1//2 * p == 0.5 * p
true

Base.:+ — Method

+(t, p)
+(p, t)

Translate the polytope p by the vector t. Returns a new Polytope.

See also map.

Examples

julia> p = Polytope([0 0; 1 0; 0 1; 1 1]);

julia> collect(vertices(-[1, 1] + 2p))
4-element Vector{Vector{Rational{BigInt}}}:
 [-1, -1]
 [1, -1]
 [-1, 1]
 [1, 1]

Base.:+ — Method

+(p::Polytope, q::Polytope)

Minkowski sum of polytopes p and q.

Examples

Hypercubes are Minkowski sums of line segments:

julia> e1 = Polytope([-1 0; 1 0]);

julia> e2 = Polytope([0 -1; 0 1]);

julia> e1 + e2 == Polytope([-1 -1; 1 -1; -1 1; 1 1])
true

Base.map — Method

map(f, p::Polytope)

Create a new Polytope whose vertices are the images of the vertices of p under the function f.

Examples

julia> p = Polytope([0 0 0; 1 0 0; 0 2 0; 0 0 3]);

julia> q = map(x -> x[1:2], p);

julia> collect(vertices(q))
3-element Vector{Vector{Rational{BigInt}}}:
 [0, 0]
 [1, 0]
 [0, 2]

projects p onto the first two coordinates.

julia> q = map(x -> x .+ 1, p);

julia> collect(vertices(q))
4-element Vector{Vector{Rational{BigInt}}}:
 [1, 1, 1]
 [2, 1, 1]
 [1, 3, 1]
 [1, 1, 4]

translates p by the vector [1, 1, 1].

Spindles.Polytopes.polarize — Method

polarize(p::Polytope)

Compute the polar dual of the polytope p.

If $V$ denotes the set of vertices of p, then the polar dual of p is defined as

\[\{ x \colon v^\top x \le 1 \text{ for all } v \in V \}\]

which is a polytope if and only if p contains the origin in its interior. If this is not the case, polarize throws an error.

Examples

Hypercubes and cross-polytopes are dual to each other:

julia> cube(3) == polarize(crosspolytope(3))
true

julia> crosspolytope(4) == polarize(cube(4))
true

Note that polar duality is an involution:

julia> p = cube(3);

julia> p == polarize(polarize(p))
true

The following polytope (a simplex) has a vertex at the origin. Therefore, polarize throws an error:

julia> p = Polytope([0 0; 1 0; 0 1]);

julia> polarize(p)
ERROR: got a polytope that does not contain the origin in its interior
[...]

However, we may make the origin an interior point by taking an arbitrary interior point x of p and shifting the polytope by -x. Here we use the centroid of the simplex for x:

julia> x = sum(vertices(p)) / nvertices(p)
2-element Vector{Rational{BigInt}}:
 1//3
 1//3

julia> polarize(-x + p)
Polytope{Rational{BigInt}}