Operators

Spindles.jl supports common operations on polytopes.

Set-theoretic operators

Base.:== — Method

==(p::Polytope, q::Polytope) -> Bool

Check whether polytopes p and q have the same set of vertices.

Examples

julia> p = Polytope([1 0; 0 1]);

julia> p == p
true

julia> p == Polytope(collect(vertices(p)))
true

julia> p == Polytope([1 0; 0 1; 0 1; 1//2 1//2])
true

julia> p == Polytope([1.0 -0.0; 0 1])
true

julia> p == Polytope([0.999999 0; 0 1])
false

Base.in — Method

in(x, p::Polytope) -> Bool
∈(x, p::Polytope) -> Bool

Check whether the vector x is in the polytope p.

Examples

julia> p = Polytope([0 0; 1 0; 0 1]);

julia> [1, 1] in p
false

julia> (sum(vertices(p)) / nvertices(p)) ∈ p
true

Base.intersect — Method

intersect(p::Polytope, polytopes...)
∩(p::Polytope, polytopes...)

Intersection of p and all polytopes in polytopes. Returns a new Polytope.

See also union.

Examples

julia> p = Polytope([[-1, 0], [1, 0]]) ∩ Polytope([[0, -1], [0, 1]]);

julia> collect(vertices(p))
1-element Vector{Vector{Rational{BigInt}}}:
 [0, 0]

julia> p = Polytope([0 0; 1 0; 0 1]);

julia> nvertices(
           intersect((Polytope([v]) for v in vertices(p))...)
       )
0

Base.isdisjoint — Method

isdisjoint(p::Polytope, q::Polytope)

Check whether p and q are disjoint. Equivalent to isempty(p ∩ q).

See also intersect, isempty.

Examples

julia> isdisjoint(Polytope([[-1, 0], [1, 0]]), Polytope([[0, -1], [0, 1]]))
false

julia> isdisjoint(Polytope([[1, 0]]), Polytope([[0, 1]]))
true

Base.isempty — Method

isempty(p::Polytope) -> Bool

Check whether p is the empty polytope $\emptyset$.

Example

julia> isempty(Polytope([[1, 0], [0, 1]]))
false

julia> isempty(Polytope([[1, 0]]) ∩ Polytope([[0, 1]]))
true

Base.issubset — Method

issubset(p::Polytope, q::Polytope) -> Bool
⊆(p::Polytope, q::Polytope) -> Bool
⊇(q::Polytope, p::Polytope) -> Bool

Check whether p is contained in q.

Examples

julia> p = Polytope([0 0; 1 0; 0 1]);

julia> issubset(p, p)
true

julia> q = Polytope([0 0; 1 0; 0 1; 1 1]);

julia> p ⊆ q
true

julia> p ⊇ q
false

Base.union — Method

union(p::Polytope, polytopes...)
∪(p::Polytope, polytopes...)

Union of p and all polytopes in polytopes. Returns a new Polytope.

See also intersect.

Examples

julia> p = Polytope([[-1, 0], [1, 0]]) ∪ Polytope([[0, -1], [0, 1]]);

julia> collect(vertices(p))
4-element Vector{Vector{Rational{BigInt}}}:
 [-1, 0]
 [1, 0]
 [0, -1]
 [0, 1]

julia> p == union((Polytope([v]) for v in vertices(p))...)
true

Sums and products of polytopes

Base.:+ — Method

+(p::Polytope, q::Polytope)

Minkowski sum of polytopes p and q.

Examples

Hypercubes are Minkowski sums of line segments:

julia> p = Polytope([-1 0; 1 0]) + Polytope([0 -1; 0 1]);

julia> collect(vertices(p))
4-element Vector{Vector{Rational{BigInt}}}:
 [-1, -1]
 [-1, 1]
 [1, -1]
 [1, 1]

Base.sum — Function

sum(ps)

Minkowski sum of the collection of Polytopes ps.

Examples

julia> p = sum([Polytope([[0, 0], Int.(1:2 .== i)]) for i=1:2]);

julia> collect(vertices(p))
4-element Vector{Vector{Rational{BigInt}}}:
 [0, 0]
 [0, 1]
 [1, 0]
 [1, 1]

Base.:* — Method

*(p::Polytope, q::Polytope)

Cartesian product of polytopes p and q.

Examples

A cube is the product of line segments:

julia> p = Polytope([[0], [1]]);

julia> collect(vertices(p * p * p))
8-element Vector{Vector{Rational{BigInt}}}:
 [1, 1, 1]
 [1, 1, 0]
 [1, 0, 1]
 [1, 0, 0]
 [0, 1, 1]
 [0, 1, 0]
 [0, 0, 1]
 [0, 0, 0]

Transformations

Base.:+ — Method

+(t, p)
+(p, t)

Translate the polytope p by the vector t. Returns a new Polytope.

See also -, map.

Examples

julia> p = Polytope([0 0; 1 0; 0 1; 1 1]);

julia> collect(vertices(-[1, 1] + 2p))
4-element Vector{Vector{Rational{BigInt}}}:
 [-1, -1]
 [1, -1]
 [-1, 1]
 [1, 1]

Base.:- — Method

-(p, t)

Equivalent to +(p, -t).

Examples

julia> p = Polytope([0 0; 1 0; 0 1; 1 1]);

julia> collect(vertices(2p - [1, 1]))
4-element Vector{Vector{Rational{BigInt}}}:
 [-1, -1]
 [1, -1]
 [-1, 1]
 [1, 1]

Base.:* — Method

*(δ, p::Polytope)
*(p::Polytope, δ)

Rescale p by the scalar factor δ. Returns a new Polytope.

As with standard scalar multiplication, the shorthands δp and -p (if δ == 1) also work here.

See also -, /, //, map.

Examples

julia> p = Polytope([[0, 0], [1, 0], [0, 1]])
Polytope{Rational{BigInt}} in 2-space

julia> collect(vertices(p * 2))
3-element Vector{Vector{Rational{BigInt}}}:
 [0, 0]
 [2, 0]
 [0, 2]

julia> 2p == p + p
true

julia> collect(vertices(-p))
3-element Vector{Vector{Rational{BigInt}}}:
 [0, 0]
 [-1, 0]
 [0, -1]

Note that the element type of the rescaled polytope may be different from that of p:

julia> 1//2 * p
Polytope{Rational{BigInt}} in 2-space

julia> 0.5 * p
Polytope{BigFloat} in 2-space

julia> 1//2 * p == 0.5 * p
true

Base.:/ — Function

/(p::Polytope, δ)

Equivalent to *(p, 1/δ).

If the scalar δ is an integer or a rational number and the element type of p is a subtype of Integer or Rational, use // to obtain a rational polytope again.

See also //.

Base.:// — Function

//(p::Polytope, δ)

Equivalent to *(p, 1//δ) where both the element type of p and the type of δ must be subtypes of Integer or Rational.

See also /.

Examples

julia> p = Polytope([[0, 0], [1, 0], [0, 1//1]])
Polytope{Rational{Int64}} in 2-space

julia> p / 1
Polytope{Float64} in 2-space

julia> p // 1
Polytope{Rational{Int64}} in 2-space

julia> (p / 1) // 1
ERROR: MethodError: no method matching //(::Polytope{Float64}, ::Int64)
[...]

Base.:- — Method

-(p::Polytope)

Equivalent to *(-1, p).

To define more complex mappings of the set of vertices, use map. See below for an example of how projections can be defined using map. More examples can be found in this tutorial.

Base.map — Function

map(f, p::Polytope)

Create a new Polytope whose vertices are the images of the vertices of p under the function f.

Examples

julia> p = Polytope([0 0 0; 1 0 0; 0 2 0; 0 0 3]);

julia> q = map(x -> x[1:2], p);

julia> collect(vertices(q))
3-element Vector{Vector{Rational{BigInt}}}:
 [0, 0]
 [1, 0]
 [0, 2]

projects p onto the first two coordinates.

julia> q = map(x -> x .+ 1, p);

julia> collect(vertices(q))
4-element Vector{Vector{Rational{BigInt}}}:
 [1, 1, 1]
 [2, 1, 1]
 [1, 3, 1]
 [1, 1, 4]

translates p by the all-ones vector [1, 1, 1].

Polarization

Spindles.Polytopes.polarize — Function

polarize(p::Polytope)

Compute the polar dual of the polytope p.

If $V$ denotes the set of vertices of p, then the polar dual of p is

\[\{ x \colon v^\top x \le 1 \text{ for all } v \in V \}\]

which is a polytope if and only if p contains the origin in its interior. If this is not the case, polarize throws an error.

Examples

Hypercubes and cross-polytopes are dual to each other:

julia> cube(3) == polarize(crosspolytope(3))
true

julia> crosspolytope(4) == polarize(cube(4))
true

Note that polar duality is an involution:

julia> p = cube(3);

julia> p == polarize(polarize(p))
true

The following polytope (a simplex) has a vertex at the origin. Therefore, polarize throws an error:

julia> p = Polytope([0 0; 1 0; 0 1]);

julia> polarize(p)
ERROR: got a polytope that does not contain the origin in its interior
[...]

However, we may make the origin an interior point by taking an arbitrary interior point x of p and shifting the polytope by -x. Here we use the centroid of the simplex for x:

julia> x = sum(vertices(p)) / nvertices(p)
2-element Vector{Rational{BigInt}}:
 1//3
 1//3

julia> polarize(p - x)
Polytope{Rational{BigInt}} in 2-space