Polytopes

The main type defined by Spindles.jl is the type Polytope that represents a polytope. Objects of this type are constructed either from a V-representation (a list of points whose convex hull is the polytope) or from an H-representation (a system of linear inequalities whose set of solutions is the polytope):

Spindles.Polytopes.Polytope — Type

Polytope(V::AbstractVector{AbstractVector})

Create a polytope from the convex hull of the collection of points V.

Polytope(V::AbstractMatrix)

Create a polytope from the convex hull of the rows of V.

Polytope(A::AbstractMatrix, b::AbstractVector)

Create a polytope from its H-representation $Ax \le b$. If the polyhedron defined by $Ax \le b$ is unbounded, throw an error.

Examples

julia> p = Polytope([[0, 0], [1, 0], [0, 1], [1, 1]])
Polytope{Rational{BigInt}} in 2-space

creates the polytope with vertices $(0,0),(1,0),(0,1)$, and $(1,1)$ (the two-dimensional 0/1 cube). This is the same as

julia> q = Polytope([0 0; 1 0; 0 1; 1 1])
Polytope{Rational{BigInt}} in 2-space

julia> p == q
true

Note

Two Polytopes are considered equal by the == operator if and only if they have the same set of vertices.

An equivalent description of the 2D polytope p is in terms of the following system of linear inequalities:

\[\begin{aligned} 0 \le x_1 &\le 1 \\ 0 \le x_2 &\le 1 \end{aligned}\]

This translates to

julia> A = [ -1   0
              1   0
              0  -1
              0   1 ];

julia> b = [0, 1, 0, 1];

julia> r = Polytope(A, b)
Polytope{Rational{BigInt}} in 2-space

julia> p == r
true

Unlike for the first two constructors above, there is no guarantee that a polyhedron defined by a general system $Ax \le b$ is bounded (and, hence, a polytope). Indeed, if we drop any of the four inequalities above – say the last one –, this property is lost:

julia> Polytope(A[1:3,:], b[1:3])
ERROR: ArgumentError: got an unbounded polyhedron
[...]

Even though all examples so far only featured minimal descriptions of the two-dimensional 0/1 cube, a Polytope object can be created from any description of the polytope, not necessarily a minimal one. In particular, the list of points whose convex hull is the polytope can include non-vertices. Likewise, redundant inequalities and implicit equations in a system of linear inequalities are permitted. Such redundancy can be detected, see Representations.

Type parameter

Polytope is a parametric type. Namely, the precise type of each of the three objects constructed above is Polytope{Rational{BigInt}}, where the parameter Rational{BigInt} is called the element type and is inferred from the type of the data passed to the constructor. For example, Rational{BigInt} indicates that Polytope uses exact rational arithmetic to store and manipulate the coefficients in a V- or H-representation. For integer data, this is the default choice.

Suppose that we change some of the input data above to floating-point numbers, such as

julia> Polytope([-1.0 0.0; 1 0; 0 -1; 0 1], [0, 1, 0, 1])
Polytope{Float64} in 2-space

Then we get a different element type, namely Float64.

The type of arithmetic used to represent and manipulate the Polytope object can also be set using an optional constructor argument that is borrowed from Polyhedra.jl (and that may be passed to any of the Polytope constructors):

julia> using Polyhedra: DefaultLibrary

julia> Polytope([-1.0 0.0; 1 0; 0 -1; 0 1], [0, 1, 0, 1], DefaultLibrary{Rational{BigInt}}())
Polytope{Rational{BigInt}} in 2-space

This argument now specifies a library for polyhedral computations (the "backend" of Polyhedra.jl) that is to be used for the internal representation of the Polytope object. In this case, we chose the default library implemented in Polyhedra.jl but forced rational data instead of Floats.

See also the JuliaPolyhedra website for a list of all supported libraries. For example, to use CDDLib with exact rational arithmetic, do

import CDDLib
Polytope(A, b, CDDLib.Library(:exact))