Representations

We saw that a Polytope may be created from either a V- or a H-representation. Spindles.jl provides functions to retrieve both types of representations for a given polytope, and to detect redundancy.

Vertices

A polytope is the convex hull of its vertices.

Spindles.Polytopes.vertices — Function

vertices(p::Polytope)

Return an iterator over the vertices of the polytope p.

See also nvertices.

Examples

julia> p = Polytope([    1     0
                         0     1
                         0     1
                      1//2  1//2 ]);

julia> collect(vertices(p))
2-element Vector{Vector{Rational{Int64}}}:
 [1, 0]
 [0, 1]

Spindles.Polytopes.nvertices — Function

nvertices(p::Polytope)

Return the number of vertices of p.

See also vertices.

The function vertices is also used to test two polytopes for equality with the == operator, see Operators.

Linear systems and redundancy

Each vertex returned by Spindles.Polytopes.vertices is a vector in the ambient space of the polytope. The dimension of this space is given by

Spindles.Polytopes.ambientdim — Function

ambientdim(p::Polytope)

Return the dimension of the ambient space of the polytope p.

See also dim.

Examples

julia> ambientdim(Polytope([[1, 0], [0, 1]]))
2

Note that this may be different from the actual dimension of a polytope (see dim) when the polytope is contained in a proper affine subspace of the ambient space. The smallest such subspace is the affine hull of the polytope. Every H-representation of a polytope must include constraints that define its affine hull. They may be included as explicit equality constraints, or as so-called implicit equations. These are inequalities that are satisfied at equality by all points in the polytope. Both types of equality constraints can be collected using affinehull.

Spindles.Polytopes.affinehull — Function

affinehull(p::Polytope; remove_rescaled=false)

Return a tuple (B, d) of a matrix B and a vector d such that the system of linear equations $Bx = d$ defines the affine hull of p.

Note that this system is not necessarily minimal. If remove_rescaled is true, a sufficient subsystem is returned for which no two equations are scalar multiples of one another.

See also inequalities, facets.

Examples

The polytope given by

\[\begin{aligned} 0 \le x_1 &\le 1 \\ x_2 &= 0 \end{aligned}\]

can be modeled by replacing the equality constraint with two inequalities $\pm x_2 \le 0$ as follows:

julia> A = [-1 0; 1 0; 0 -1; 0 1]; b = [0, 1, 0, 0];

julia> p = Polytope(A, b);

julia> affinehull(p)
(Rational{BigInt}[0 -1; 0 1], Rational{BigInt}[0, 0])

julia> affinehull(p; remove_rescaled=true)
(Rational{BigInt}[0 -1], Rational{BigInt}[0])

To obtain a complete H-representation, we further need inequalities to describe the facets of p.

Spindles.Polytopes.inequalities — Function

inequalities(p::Polytope)

Return a tuple (A, b) of a matrix A and a vector b such that p is given by all points $x$ in the affine hull of p that satisfy $Ax \le b$.

If p was created from an H-representation and is full-dimensional, this will return the same system that p was created from.

See also ineqindices, facets, affinehull.

Examples

julia> A = [-1 0; 0 -1; 2 3]; b = [0, 0, 1];

julia> inequalities(Polytope(A, b)) == (A, b)
true

Spindles.Polytopes.ineqindices — Function

ineqindices(p::Polytope)

Return the valid range of indices for the rows of the coefficient matrix A returned by inequalities.

See also inequalities, Base.axes.

Examples

julia> ineqindices(Polytope([-1 0; 0 -1; 2 3], [0, 0, 1]))
Base.OneTo(3)

Note that the system returned by inequalities is not necessarily minimal, i.e., it may include redundant inequalities. A redundant inequality is one whose deletion leaves the polytope unchanged. A sufficient subset of inequalities from the given description returned by inequalities can be extracted using the function facets.

Spindles.Polytopes.facets — Function

facets(p::Polytope)

Return a tuple (A, b) such that the system $Ax \le b$ is a minimal system of facet-definining inequalities for p.

Note that this system need not be unique. If p was constructed from an H-representation for which multiple inequalities define the same facet, the facet-defining inequality with the smallest index is selected.

See also nfacets, inequalities.

Examples

julia> A = [ -1   0
              1   0
              0  -1
              0   1 ];

julia> b = [0, 1, 0, 1];

julia> p = Polytope(A, b);

creates the polytope defined by the system

\[\begin{aligned} 0 \le x_1 &\le 1 \\ 0 \le x_2 &\le 1 \end{aligned}\]

which is minimal:

julia> facets(p) == (A, b)
true

The following inequalities are also valid for p:

\[\begin{aligned} 3 x_1 &\le 3 \\ x_1 + x_2 &\le 2 \end{aligned}\]

Adding them to the original (irredundant) system introduces redundancy:

julia> q = Polytope([3 0; 1 1; A], [3; 2; b]);

julia> B, d = facets(q);

julia> B
4×2 Matrix{Rational{BigInt}}:
  3   0
 -1   0
  0  -1
  0   1

julia> d
4-element Vector{Rational{BigInt}}:
 3
 0
 0
 1

julia> p == q == Polytope(B, d)
true

Spindles.Polytopes.nfacets — Function

nfacets(p::Polytope)

Return the number of facets of p. Equivalent to nfacesofdim(p, dim(p) - 1) but has a different implementation.

See also nfacesofdim, dim, facets.

Incidence

Spindles.Polytopes.incidentvertices — Function

incidentvertices(p::Polytope, indices::AbstractVector{Int})

Return the indices of all vertices of the polytope p for which each inequality in the collection indices is tight (i.e., satisfied at equality). Here, index i refers to the ith inequality in inequalities.

If indices is empty, this is the same as collect(1:nvertices(p)).

See also nvertices, inequalities, tightinequalities.

Examples

julia> p = Polytope([-1 0; 0 -1; 2 3], [0, 0, 1]);

julia> collect(vertices(p))
3-element Vector{Vector{Rational{BigInt}}}:
 [1//2, 0]
 [0, 1//3]
 [0, 0]

julia> inequalities(p)
(Rational{BigInt}[-1 0; 0 -1; 2 3], Rational{BigInt}[0, 0, 1])

julia> incidentvertices(p, [3])
2-element Vector{Int64}:
 1
 2

Spindles.Polytopes.tightinequalities — Function

tightinequalities(p::Polytope, indices)

Return the indices of all inequalities that are tight (i.e., satisfied at equality) for each vertex in the collection indices.

If indices is empty, this is the same as collect(ineqindices(p)).

See also ineqindices, incidentvertices.

Examples

julia> p = Polytope([-1 0; 0 -1; 2 3], [0, 0, 1]);

julia> collect(vertices(p))
3-element Vector{Vector{Rational{BigInt}}}:
 [1//2, 0]
 [0, 1//3]
 [0, 0]

julia> inequalities(p)
(Rational{BigInt}[-1 0; 0 -1; 2 3], Rational{BigInt}[0, 0, 1])

julia> tightinequalities(p, [1])
2-element Vector{Int64}:
 2
 3