Spindles

Spindles are special polytopes for which each facet is incident to exactly one of two special vertices.

Spindles.Polytopes.apices — Function

apices(p::Polytope [, apex::Int]; checkredund=true)

Check whether p is a spindle, i.e., whether there is a pair of vertices (the apices) for which each facet of p is incident to exactly one of them. If p has a pair of apices, return their indices; otherwise return nothing.

The optional argument apex specifies the index of a vertex that is to be taken as one of the apices.

Keywords

checkredund::Bool: If true (default), first detect all inequalities in the description of p that are not facet-defining for p. If p was created from an inequality description that is known to be minimal, this step may be skipped by setting checkredund to false.

Note

In the presence of redundant inequalities, disabling checkredund can only produce false negatives: Whenever apices returns a pair of vertices, they are guaranteed to be apices of p. The converse, however, is not necessarily true. A spindle may not be detected as such if its inequality description contains inequalities that are not facet-defining.

Examples

julia> p = Polytope([1 0; 0 1; -1 0; 0 -1], [1, 1, 1, 1]);

julia> vertices(p)
4-element iterator of Vector{Rational{BigInt}}:
 Rational{BigInt}[-1, -1]
 Rational{BigInt}[1, -1]
 Rational{BigInt}[-1, 1]
 Rational{BigInt}[1, 1]

julia> apices(p)
2-element Vector{Int64}:
 1
 4

julia> apices(p, 2)
2-element Vector{Int64}:
 2
 3