Representation

Spindles — Module

Main module in Spindles.jl.

Exports

Spindle

apices

vertices

nvertices

...

The main type defined by Spindles.jl is

Spindles.Spindle — Type

Spindle

Main type that represents a spindle...

Objects of this type may be created in one of two possible ways:

Spindles.Spindle — Method

Spindle(A, b [,lib])

Create a spindle from its inequality description with coefficient matrix A and right-hand side vector b.

The optional argument lib allows to specify a library for polyhedral computations that implements the interface of Polyhedra. A list of all supported libraries can be found on the JuliaPolyhedra website.

If lib is not specified, use the default library implemented in Polyhedra (see the Polyhedra documentation).

Warning

The current version supports only full-dimensional polytopes given by irredundant inequality descriptions. In particular, A and b are not checked for the presence of redundant rows or implicity equations, or whether they define a polytope at all.

Spindle(P::Polyhedron)

Create a spindle directly from a Polyhedron P.

Vertices and apices

Spindles.vertices — Function

vertices(s)

Returns an iterator over the vertices of the spindle s.

Spindles.nvertices — Function

nvertices(s)

Count the vertices of s.

Spindles.apices — Function

apices(s [,apex])

Compute a pair of vertices (the apices) such that each facet of s is incident to exactly one of them, or throw an error if no such pair exists.

If additionally given the index of a vertex apex, try to find a vertex distinct from apex such that the two vertices are apices of s.

Spindles.nfacets — Function

nfacets(s)

Incidence

Spindles.incidentvertices — Function

incidentvertices(s, facets)

List the indices of all vertices of the spindle s that are incident with facets. Returns an iterator.