Faces

Enumeration

Spindles.graph — Function

graph(s::Spindle)

Return the graph (aka 1-skeleton) of s, which is a simple undirected graph of type Graphs.SimpleGraphs.SimpleGraph.

Spindles.jl implements an algorithm for enumerating all faces of a given dimension. On large inputs, this algorithm works best for near-simple spindles (with few degenerate vertices).

Spindles.facesofdim — Function

facesofdim(s::Spindle, k::Int)

Enumerate all faces of dimension k of the spindle s. Each face is given by a list of the indices of its incident halfspaces/facets.

Note

The empty face $\emptyset$ (which is the unique face of dimension -1 by convention) is given by the list of all facet/halfspace indices, as the intersection of all facets of a polytope is empty.

Difference to Polyhedra.jl

The index of a halfspace/facet is its index in the collection of halfspaces of the polyhedron underlying s, where hyperplanes are ignored. This is different from the way that indices are treated in Polyhedra, where hyperplanes and halfspaces share the same set of indices.

The algorithm proceeds recursively and computes faces bottom-up, starting from the vertices.

Note

Results are cached internally in the Spindle object s. Therefore, subsequent calls to facesofdim(s, l) for any $l \le k$ do not cost anything.

See also nfacesofdim.

Spindles.nfacesofdim — Function

nfacesofdim(s::Spindle, k::Int)

Count the k-dimensional faces of the spindle s. Shorthand for length(facesofdim(s, k)).

See also facesofdim.

Length of paths

Spindles.dist — Function

dist(s::Spindle, apex::Int, v::Int)

Compute the distance between apex and vertex v in the graph of spindle s.

Note

Results are affected by resetting the apices with setapex!.

Examples

julia> square = Spindle([1 0; 0 1; -1 0; 0 -1], [1, 1, 1, 1]);

julia> apx1, apx2 = setapex!(square, 1)
2-element Vector{Int64}:
 1
 4

julia> dist(square, apx1, 4)
2

julia> apx1, apx2 = setapex!(square, 2)
2-element Vector{Int64}:
 2
 3

julia> dist(square, apx1, 4)
1

Good faces

Spindles.FaceState — Type

FaceState

Fields

good::Bool

facets: all incident facets

edges

vsets

Spindles.isgood2face — Function

isgood2face(s::Spindle, facets)

Test the face defined by facets for being a good 2-face of the spindle s. Return a FaceState.

See this tutorial for an informal explanation of what it means for a 2-face to be good.