Faces and graphs

Enumeration

Spindles.jl implements an algorithm for enumerating all faces of a given dimension. The algorithm is optimized for near-simple polytopes, i.e., polytopes with few degenerate vertices. It is inspired by an algorithm for computing the graph of a polytope described here (see also the paper).

Spindles.facesofdim — Function

facesofdim(p::Polytope, k::Int)

Enumerate all faces of dimension k of the polytope p. Each face is given by a list of the indices of its incident halfspaces.

Note here that the empty face $\emptyset$ (which is the unique face of dimension −1 by convention) is given by the list of all halfspace indices, as the intersection of all facets of a polytope is empty.

Difference from Polyhedra.jl

The index of a halfspace is the index of the corresponding inequality in the linear description of p, where (explicitly given) equality constraints (defining hyperplanes) are ignored. This is different from the way that indices are treated in Polyhedra.jl, where hyperplanes and halfspaces share the same set of indices.

Note

The algorithm proceeds recursively and computes faces bottom-up, starting from the vertices. The results are cached internally in the Polytope object p. Therefore, subsequent calls to facesofdim(p, l) for any $l \le k$ do not cost anything.

See also nfacesofdim, dim.

Spindles.nfacesofdim — Function

nfacesofdim(p::Polytope, k::Int)

Count the k-dimensional faces of the polytope p. Shorthand for length(facesofdim(p, k)).

See also nfacets, facesofdim, dim.

Examples

julia> V = [(i >> j) & 1 for i=0:7, j=0:2]
8×3 Matrix{Int64}:
 0  0  0
 1  0  0
 0  1  0
 1  1  0
 0  0  1
 1  0  1
 0  1  1
 1  1  1

julia> cube = Polytope(V);

julia> nfacesofdim.(cube, -1:3)
5-element Vector{Int64}:
  1
  8
 12
  6
  1

For example, calling facesofdim(p, 1) lists the incident halfspaces of each edge of the polytope p. These edges, given as pairs of adjacent vertices, define the graph (or 1-skeleton) of p. The graph can also be retrieved directly:

Spindles.graph — Function

graph(p::Polytope)

Return the graph of the polytope p, which is a simple undirected graph of type Graphs.SimpleGraphs.SimpleGraph.

Length of paths

Spindles.dist — Function

dist(p::Polytope, u::Int, v::Int)

Compute the distance between vertices u and v in the graph of p.

Examples

julia> p = Polytope([0 0; 0 1; 1 1; 1 0]);

julia> vertices(p)
4-element iterator of Vector{Rational{BigInt}}:
 Rational{BigInt}[0, 0]
 Rational{BigInt}[0, 1]
 Rational{BigInt}[1, 1]
 Rational{BigInt}[1, 0]

julia> dist(p, 1, 3)
2

Good faces

Spindles.FaceState — Type

FaceState

Fields

good::Bool

indices: indices of all incident halfspaces

edges

vsets

Spindles.isgood2face — Function

isgood2face(p::Polytope, indices, src, dst)

Test the face defined by indices for being a good 2-face of the polytope p with respect to the two vertices src and dst. Return a FaceState.

See this tutorial for an informal explanation of what it means for a 2-face to be good.