Reading and writing files

Spindles.writeineq — Function

writeineq(fname, A::AbstractMatrix, b::AbstractVector [, rowlabels];
    labels=true, comments=[], comment_char='#'
)

Write the coefficient matrix A and the vector of right-hand sides b of a system of linear inequalities to the file fname.

The file format is [rowlabels b -A] where rowlabels is a vector of strings, or just [b -A] (without a label column) if labels is false.

If no rowlabels are provided and labels is true, use the row indices as default labels.

See also readineq.

Keywords

labels: If true (default), print a first column with row labels as given by the argument rowlabels. If rowlabels is unspecified, each row is labeled by its index (ranging from 1 to the number of rows of A).

comments: A vector of strings. Each element in comments will be printed on its own line, following a comment_char and a whitespace. Possible internal line breaks are ignored. The data A and b is printed below the last line of comments.

comment_char: An AbstractChar that indicates the start of a comment line (default is '#').

Examples

julia> A = [1 0; 0 1; -1 0; 0 -1];

julia> b = [1, 1, 1, 1];

julia> labels = ["α", "β", "γ", "δ"];

julia> writeineq("ineq.txt", A, b, labels; comments=["A nice polytope"])

julia> print(read("ineq.txt", String))
# A nice polytope
α  1  -1   0
β  1   0  -1
γ  1   1   0
δ  1   0   1

julia> A, b, labels = readineq("ineq.txt", Int);

julia> A
4×2 Matrix{Rational{Int64}}:
  1   0
  0   1
 -1   0
  0  -1

julia> b
4-element Vector{Rational{Int64}}:
 1
 1
 1
 1

julia> labels
4-element Vector{String}:
 "α"
 "β"
 "γ"
 "δ"

julia> rm("ineq.txt")

Spindles.readineq — Function

readineq(fname, T; labels=true, comment_char='#')

Read A, b, labels from file fname. The element type of A and b is Rational{T}.

Lines starting with a comment_char character and all characters on a line following a comment_char are ignored.

See also writeineq.

Keywords

labels:

comment_char: