Minimal representations

We saw that a Polytope may be created from either a V- or a H-representation. However, these representations need not be minimal. Which points or inequalities from the given description suffice can be detected with the following functions.

Vertices

To express a polytope as the convex hull of some finite set of points, the set of its vertices suffices.

Spindles.vertices — Function

vertices(p::Polytope)

Return an iterator over the vertices of the polytope p.

Examples

julia> p = Polytope([1 0; 0 1; 0 1; 1//2 1//2]);

julia> collect(vertices(p))
2-element Vector{Vector{Rational{Int64}}}:
 [1, 0]
 [0, 1]

Spindles.nvertices — Function

nvertices(p::Polytope)

Count the vertices of p.

The function vertices is also used to test two polytopes for equality:

Base.:== — Function

==(p::Polytope, q::Polytope)

Check whether the sets of vertices of p and q are identical.

Examples

julia> Polytope([1 0; 0 1]) == Polytope([1 0; 0 1; 0 1; 1//2 1//2])
true

Redundancy and implicit equations

A redundant inequality is one whose deletion leaves the polytope unchanged. If an inequality is satisfied at equality by all points in the polytope, it is said to be an implicit equation. Thus, the inequalities in a given description of a polytope may be partitioned into three sets:

a minimal set of facet-defining inequalities (which may not be unique!),

a (possibly empty) set of implicit equations contained in the inequality system, and

all remaining inequalities (that may safely be deleted).

This partition can be computed using the functions facets and impliciteqs, which return the first two sets of the partition.

Spindles.facets — Function

facets(p::Polytope)

Return a minimal subset of inequality indices that contains one inequality for each facet of p.

If multiple inequalities in the description of p define the same facet, the one with the smallest index is selected.

See also nfacets, impliciteqs.

Examples

julia> A = [-1 0; 1 0; 0 -1; 0 1]; b = [0, 1, 0, 1];

julia> p = Polytope(A, b);

creates the polytope defined by the system

\[\begin{aligned} 0 \le x_1 &\le 1 \\ 0 \le x_2 &\le 1 \end{aligned}\]

The following inequalities are also valid for p:

\[\begin{aligned} 3 x_1 &\le 3 \\ x_1 + x_2 &\le 2 \end{aligned}\]

Adding them to the original (irredundant) system introduces redundancy:

julia> B = [3 0; 1 1; A]; d = [3; 2; b];

julia> q = Polytope(B, d);

julia> f = facets(q)
4-element Vector{Int64}:
 1
 3
 5
 6

julia> r = Polytope(B[f,:], d[f]);

julia> p == q == r
true

Spindles.nfacets — Function

nfacets(p::Polytope)

Count the facets of p. Shorthand for length(facets(p)).

See also nfacesofdim, facets.

Spindles.impliciteqs — Function

impliciteqs(p::Polytope)

Return the indices of all inequalities that are implicit equations, i.e., that are satisfied at equality for each point in p.

See also facets.

Examples

The polytope given by

\[\begin{aligned} 0 \le x_1 &\le 1 \\ x_2 &= 0 \end{aligned}\]

can be modeled by replacing the equality constraint with two inequalities $\pm x_2 \le 0$ as follows:

julia> A = [-1 0; 1 0; 0 -1; 0 1]; b = [0, 1, 0, 0];

julia> p = Polytope(A, b);

julia> impliciteqs(p)
2-element Vector{Int64}:
 3
 4

Incidence

Spindles.incidentvertices — Function

incidentvertices(p::Polytope, indices::AbstractVector{Int})

List the indices of all vertices of the polytope p for which each inequality in indices is tight.

If indices is empty, this is the same as collect(1:nvertices(p)).