Representation

Spindles — Module

Main module. The public API is listed here.

The main type defined by Spindles.jl is

Spindles.Spindle — Type

Spindle{T}

Main type that represents a spindle.

There are two possible ways to construct objects of this type. The first way explicitly uses Polyhedra:

Spindles.Spindle — Method

Spindle{T}(p::Polyhedra.Polyhedron{T})

Create a spindle from the polyhedron p. If p is unbounded or there are no two apices among its vertices, throw an error.

If not specified, the element type T is the element type of p.

See also Polyhedra.polyhedron.

Examples

julia> using Polyhedra: polyhedron, vrep, hrep

julia> p = polyhedron(vrep([0 0; 1 0; 0 1; 1 1]));

julia> Spindle(p)
Spindle{Rational{BigInt}}

creates the two-dimensional 0/1 cube whose vertices are $(0,0),(1,0),(0,1)$, and $(1,1)$. To construct from such a V-representation a polyhedron that can be passed to the Spindle constructor, we used Polyhedra.vrep. Equivalently, we could have created it from an inequality description (an H-representation), for example from the system of inequalities

\[\begin{aligned} 0 \le x_1 &\le 1 \\ 0 \le x_2 &\le 1 \end{aligned}\]

Using Polyhedra.hrep, this translates to

julia> p = polyhedron(hrep([-1 0; 1 0; 0 -1; 0 1], [0, 1, 0, 1]));

julia> Spindle(p)
Spindle{Rational{BigInt}}

Note that the 0/1 cube is a spindle. However, if we drop any of its vertices, this property is lost:

julia> p = polyhedron(vrep([0 0; 1 0; 0 1]));

julia> Spindle(p)
ERROR: ArgumentError: not a spindle: cannot find two apices
[...]

Similarly, trying to create a spindle from a proper subset of the inequalities in the H-representation above results in an error:

julia> p = polyhedron(hrep([-1 0; 1 0; 0 -1], [0, 1, 0]));

julia> Spindle(p)
ERROR: ArgumentError: got an unbounded polyhedron
[...]

To make spindle constructions more convenient, there is a second constructor that accepts the data of an H-representation and does not require any imports from Polyhedra:

Spindles.Spindle — Type

Spindle(A::AbstractMatrix, b::AbstractVector [, lib::Polyhedra.Library])

Create a spindle from its H-representation $Ax \le b$. The optional argument lib specifies a library for polyhedral computations (the "backend" of Polyhedra.jl) and is passed to Polyhedra.hrep. If unspecified, use the default library implemented in Polyhedra.

Info

Spindle(A, b, lib) is equivalent to Spindle(polyhedron(hrep(A, b), lib)).

See also the Polyhedra.jl documentation on libraries. A list of all supported libraries can be found on the JuliaPolyhedra website.

Examples

To use CDDLib with exact rational arithmetic, do

import CDDLib
Spindle(A, b, CDDLib.Library(:exact))

Note

If the lib argument is not specified, Spindle will infer the type of arithmetic used from the input data. This behaviour is inherited from the default library in Polyhedra. For example, changing some of the entries of the coefficient matrix in the examples above to floats produces

julia> Spindle([-1.0 0.0; 1 0; 0 -1; 0 1], [0, 1, 0, 1])
Spindle{Float64}

as opposed to Spindle{Rational{BigInt}}.

Vertices and apices

Spindles.vertices — Function

vertices(s::Spindle)

Returns an iterator over the vertices of the spindle s.

Spindles.nvertices — Function

nvertices(s::Spindle)

Count the vertices of s.

Spindles.apices — Function

apices(s::Spindle)

Return the indices of a pair of vertices (the apices) of s for which each facet of s is incident to exactly one of them.

Spindles.setapex! — Function

setapex!(s::Spindle, apex::Int)

Try to find the index v of a vertex of s such that v and apex are apices of s. If successful, overwrite the apices of s, or throw an error otherwise.

Examples

julia> square = Spindle([1 0; 0 1; -1 0; 0 -1], [1, 1, 1, 1]);

julia> vertices(square)
4-element iterator of Vector{Rational{BigInt}}:
 Rational{BigInt}[-1, -1]
 Rational{BigInt}[1, -1]
 Rational{BigInt}[-1, 1]
 Rational{BigInt}[1, 1]

julia> apices(square)
2-element Vector{Int64}:
 1
 4

julia> setapex!(square, 2);

julia> apices(square)
2-element Vector{Int64}:
 2
 3

Incidence

Spindles.incidentvertices — Function

incidentvertices(s::Spindle, facets)

List the indices of all vertices of the spindle s that are incident with facets.

Note

incidentvertices(s, Int[]) is equivalent to collect(1:nvertices(s)).

Dimension

Spindles.dim — Function

dim(s::Spindle)

Compute the dimension of the spindle s.

This is done by computing the length of a maximal chain in the face poset of s, i.e., a finite sequence of faces

\[\emptyset = F_{-1} \subsetneq F_0 \subsetneq F_1 \subsetneq \dots \subsetneq F_d\]

for which d is maximal among all such sequences. Then $F_d$ must be s itself, and d is its dimension.

See also Polyhedra.dim.